+++ pkg/bastion/acl.go
@@ -1,9 +1,9 @@
-package bastion // import "moul.io/sshportal/pkg/bastion"
+package bastion // import "github.com/doovemax/sshportal/pkg/bastion"
 
 import (
 	"sort"
 
-	"moul.io/sshportal/pkg/dbmodels"
+	"github.com/doovemax/sshportal/pkg/dbmodels"
 )
 
 type byWeight []*dbmodels.ACL