+++ pkg/bastion/ssh.go
@@ -1,4 +1,4 @@
-package bastion // import "moul.io/sshportal/pkg/bastion"
+package bastion // import "github.com/doovemax/sshportal/pkg/bastion"
 
 import (
 	"bytes"
@@ -12,8 +12,8 @@ import (
 	"github.com/jinzhu/gorm"
 	"github.com/moul/ssh"
 	gossh "golang.org/x/crypto/ssh"
-	"moul.io/sshportal/pkg/crypto"
-	"moul.io/sshportal/pkg/dbmodels"
+	"github.com/doovemax/sshportal/pkg/crypto"
+	"github.com/doovemax/sshportal/pkg/dbmodels"
 )
 
 type sshportalContextKey string